refactor(sample): remove SampleAction typealias
